Shift various parts of main.js to their own files for sanity's sake
