+++ module/main.js
@@ -1,4 +1,4 @@
-/* global CONFIG, Handlebars, Hooks, Actors, ActorSheet, ChatMessage, Dialog, Items, ItemSheet, Macro, game, ui, renderTemplate, getProperty */
+/* global CONFIG, Hooks, Actors, ActorSheet, ChatMessage, Dialog, Items, ItemSheet, Macro, game, ui, renderTemplate, getProperty */
 
 // Import Modules
 import { preloadHandlebarsTemplates } from './templates.js'
@@ -28,84 +28,16 @@ import {
   watchPool2Filters,
   prepareCustomRollButton
 } from './dice/dicebox.js'
+import { loadDiceSoNice } from './dice-so-nice.js'
+import { loadHelpers } from './helpers.js'
+import { loadSettings } from './settings.js'
 
 const OWNED_PERMISSION = 3
 
 Hooks.once('init', async function () {
   console.log('Initializing Schrecknet...')
 
-  game.settings.register('vtm5e', 'worldVersion', {
-    name: 'World Version',
-    hint: 'Automatically upgrades data when the system.json is upgraded.',
-    scope: 'world',
-    config: true,
-    default: '1.5',
-    type: String
-  })
-
-  game.settings.register('vtm5e', 'darkTheme', {
-    name: 'Dark Theme',
-    hint: 'Display sheets using a darker theme on a per-user basis.',
-    scope: 'client',
-    config: true,
-    default: false,
-    type: Boolean
-  })
-
-  game.settings.register('vtm5e', 'actorBanner', {
-    name: 'Enable Character Type Banner',
-    hint: 'Display a banner at the top of actor sheets to represent the character type.',
-    scope: 'client',
-    config: true,
-    default: true,
-    type: Boolean
-  })
-
-  game.settings.register('vtm5e', 'useChatRoller', {
-    // TODO: fix Chat Roller
-    name: 'Chat Roller',
-    hint: 'Display dice roller in chat window. WARNING: Currently not working properly with Hunter changes.',
-    scope: 'world',
-    config: true,
-    default: false,
-    type: Boolean
-  })
-
-  game.settings.register('vtm5e', 'chatRollerSortAbilities', {
-    name: 'Sort Abilities in Chat Roller',
-    hint: 'Sort abilities (Attributes, Skills, Disciplines, Edges) alphabetically in the chat roller. Disable to sort in the order on the character sheet (grouping physical, social, and mental).',
-    scope: 'client',
-    config: true,
-    default: true,
-    type: Boolean
-  })
-
-  game.settings.register('vtm5e', 'automatedWillpower', {
-    name: 'Willpower Damage On Willpower Reroll',
-    hint: 'If enabled, using the Willpower Reroll (right click on a chat message) feature will deal willpower damage to the associated actor.',
-    scope: 'world',
-    config: true,
-    default: true,
-    type: Boolean
-  })
-
-  game.settings.register('vtm5e', 'automatedRouse', {
-    name: 'Increase Hunger With Rouse Checks',
-    hint: 'If enabled, rolling a rouse check and failing will automatically increase the hunger of the associated actor.',
-    scope: 'world',
-    config: true,
-    default: true,
-    type: Boolean
-  })
-
-  game.settings.register('vtm5e', 'automatedRage', {
-    name: 'Automate Rage Dice',
-    hint: 'If enabled, rolling Rage Dice or performing actions that require Rage Dice will automatically subtract Rage from the associated actor.',
-    scope: 'world',
-    config: true,
-    default: true,
-    type: Boolean
-  })
+  loadSettings()
 
   game.wod5e = {
     ActorInfo,
@@ -182,191 +114,7 @@ Hooks.once('init', async function () {
 
   preloadHandlebarsTemplates()
 
-  // If you need to add Handlebars helpers, here are a few useful examples:
-  Handlebars.registerHelper('concat', function () {
-    let outStr = ''
-    for (const arg in arguments) {
-      if (typeof arguments[arg] !== 'object') {
-        outStr += arguments[arg]
-      }
-    }
-    return outStr
-  })
-
-  Handlebars.registerHelper('ifeq', function (a, b, options) {
-    if (a === b) {
-      return options.fn(this)
-    }
-
-    return options.inverse(this)
-  })
-
-  Handlebars.registerHelper('ifnoteq', function (a, b, options) {
-    if (a !== b) {
-      return options.fn(this)
-    }
-
-    return options.inverse(this)
-  })
-
-  Handlebars.registerHelper('or', function (bool1, bool2) {
-    return bool1 || bool2
-  })
-
-  Handlebars.registerHelper('and', function (bool1, bool2) {
-    return bool1 && bool2
-  })
-
-  Handlebars.registerHelper('toLowerCase', function (str) {
-    return str.toLowerCase()
-  })
-
-  Handlebars.registerHelper('toUpperCaseFirstLetter', function (str) {
-    return str.charAt(0).toUpperCase() + str.slice(1)
-  })
-
-  const capitalize = (s) => {
-    if (typeof s !== 'string') return ''
-    return s.charAt(0).toUpperCase() + s.slice(1)
-  }
-
-  Handlebars.registerHelper('generateFeatureLabel', function (str) {
-    return 'WOD5E.'.concat(capitalize(str))
-  })
-
-  Handlebars.registerHelper('generateSkillLabel', function (str) {
-    return 'WOD5E.'.concat(str.split(' ').flatMap(word => capitalize(word)).join(''))
-  })
-
-  Handlebars.registerHelper('frenzy', function (willpowerMax, willpowerAgg, willpowerSup, humanity) {
-    // Return the result of the stain, or 1 at minimum.
-    const stainDice = Math.max((willpowerMax - willpowerAgg - willpowerSup) + Math.floor(humanity / 3), 1)
-
-    return stainDice
-  })
-
-  Handlebars.registerHelper('willpower', function (willpowerMax, willpowerAgg, willpowerSup) {
-    // Return the result of the willpower, or 1 at minimum.
-    const willpowerDice = Math.max((willpowerMax - willpowerAgg - willpowerSup), 1)
-
-    return willpowerDice
-  })
-
-  // TODO: there exist math helpers for handlebars
-  Handlebars.registerHelper('remorse', function (humanity, stain) {
-    // Return the result of the stain, or 1 at minimum.
-    const remorseDice = Math.max((10 - humanity - stain), 1)
-
-    return remorseDice
-  })
-
-  Handlebars.registerHelper('harano-test', function (harano, hauglosk) {
-    const haranoDice = Math.max((harano + hauglosk), 1)
-
-    return haranoDice
-  })
-
-  Handlebars.registerHelper('hauglosk-test', function (harano, hauglosk) {
-    const haugloskDice = Math.max((harano + hauglosk), 1)
-
-    return haugloskDice
-  })
-
-  Handlebars.registerHelper('attrIf', function (attr, value, test) {
-    if (value === undefined) return ''
-    return value === test ? attr : ''
-  })
-
-  Handlebars.registerHelper('visibleDisciplines', function (disciplines) {
-    return Object.keys(disciplines).reduce(
-      (obj, key) => {
-        if (disciplines[key].visible) {
-          obj[key] = disciplines[key]
-        }
-        return obj
-      },
-      {}
-    )
-  })
-
-  Handlebars.registerHelper('visibleEdges', function (edges) {
-    return Object.keys(edges).reduce(
-      (obj, key) => {
-        if (edges[key].visible) {
-          obj[key] = edges[key]
-        }
-        return obj
-      },
-      {}
-    )
-  })
-
-  Handlebars.registerHelper('sortAbilities', function (unordered = {}) {
-    if (!game.settings.get('vtm5e', 'chatRollerSortAbilities')) {
-      return unordered
-    }
-    return Object.keys(unordered).sort().reduce(
-      (obj, key) => {
-        obj[key] = unordered[key]
-        return obj
-      },
-      {}
-    )
-  })
-
-  Handlebars.registerHelper('numLoop', function (num, options) {
-    let ret = ''
-
-    for (let i = 0, j = num; i < j; i++) {
-      ret = ret + options.fn(i)
-    }
-
-    return ret
-  })
-
-  Handlebars.registerHelper('getDisciplineName', function (key, roll = false) {
-    const disciplines = {
-      animalism: 'WOD5E.Animalism',
-      auspex: 'WOD5E.Auspex',
-      celerity: 'WOD5E.Celerity',
-      dominate: 'WOD5E.Dominate',
-      fortitude: 'WOD5E.Fortitude',
-      obfuscate: 'WOD5E.Obfuscate',
-      potence: 'WOD5E.Potence',
-      presence: 'WOD5E.Presence',
-      protean: 'WOD5E.Protean',
-      sorcery: 'WOD5E.BloodSorcery',
-      oblivion: 'WOD5E.Oblivion',
-      alchemy: 'WOD5E.ThinBloodAlchemy',
-      rituals: 'WOD5E.Rituals',
-      ceremonies: 'WOD5E.Ceremonies'
-    }
-    if (roll) {
-      if (key === 'rituals') {
-        return disciplines.sorcery
-      } else if (key === 'ceremonies') {
-        return disciplines.oblivion
-      }
-    }
-    return disciplines[key]
-  })
-
-  Handlebars.registerHelper('getEdgeName', function (key) {
-    const edges = {
-      arsenal: 'WOD5E.Arsenal',
-      ordnance: 'WOD5E.Ordnance',
-      library: 'WOD5E.Library',
-      improvisedgear: 'WOD5E.ImprovisedGear',
-      globalaccess: 'WOD5E.GlobalAccess',
-      dronejockey: 'WOD5E.DroneJockey',
-      beastwhisperer: 'WOD5E.BeastWhisperer',
-      sensetheunnatural: 'WOD5E.SenseTheUnnatural',
-      repeltheunnatural: 'WOD5E.RepelTheUnnatural',
-      thwarttheunnatural: 'WOD5E.ThwartTheUnnatural',
-      artifact: 'WOD5E.Artifact'
-    }
-    return edges[key]
-  })
+  loadHelpers()
 })
 
 Hooks.once('ready', async function () {
@@ -375,369 +123,7 @@ Hooks.once('ready', async function () {
 })
 
 Hooks.once('diceSoNiceReady', (dice3d) => {
-  dice3d.addSystem({ id: 'vtm5e', name: 'VtM5e' }, true)
-  dice3d.addDicePreset({
-    type: 'dv',
-    labels: [
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-crit-dsn.png'
-    ],
-    colorset: 'black',
-    system: 'vtm5e'
-  })
-  dice3d.addColorset({
-    name: 'hunger',
-    description: 'V5 Hunger Dice',
-    category: 'V5',
-    foreground: '#fff',
-    background: '#6e0000',
-    texture: 'none',
-    edge: '#6e0000',
-    material: 'plastic',
-    font: 'Arial Black',
-    fontScale: {
-      d6: 1.1,
-      df: 2.2,
-      dv: 0.8,
-      dg: 0.8
-    }
-  }, 'default')
-
-  dice3d.addDicePreset({
-    type: 'dg',
-    labels: [
-      'systems/vtm5e/assets/images/bestial-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsn.png',
-      'systems/vtm5e/assets/images/red-crit-dsn.png'
-    ],
-    colorset: 'hunger',
-    system: 'vtm5e'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dh',
-    labels: [
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-crit-dsn.png'
-    ],
-    colorset: 'hunterdice',
-    system: 'vtm5e'
-  })
-
-  dice3d.addColorset({
-    name: 'hunterdice',
-    description: 'Hunter Dice',
-    category: 'V5',
-    foreground: '#000000',
-    background: '#cb650f',
-    texture: 'none',
-    edge: '#cb650f',
-    material: 'plastic',
-    font: 'Arial Black',
-    fontScale: {
-      d6: 1.1,
-      df: 2.2,
-      dv: 0.8,
-      dg: 0.8,
-      dh: 0.7
-    }
-  }, 'default')
-
-  dice3d.addColorset({
-    name: 'desperation',
-    description: 'Desperation Dice',
-    category: 'V5',
-    foreground: '#fff',
-    background: '#ee7e1f',
-    texture: 'none',
-    edge: '#ee7e1f',
-    material: 'plastic',
-    font: 'Arial Black',
-    fontScale: {
-      d6: 1.1,
-      df: 2.5,
-      dv: 0.8,
-      dg: 0.8,
-      dh: 0.7
-    }
-  }, 'default')
-
-  dice3d.addDicePreset({
-    type: 'ds',
-    labels: [
-      'systems/vtm5e/assets/images/desperation-fail-dsnc.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-crit-dsnc.png'
-    ],
-    colorset: 'black',
-    system: 'vtm5e'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dw',
-    labels: [
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-crit-dsn.png'
-    ],
-    colorset: 'werewolf',
-    system: 'vtm5e'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dr',
-    labels: [
-      'systems/vtm5e/assets/images/werewolf-brutal-fail-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-brutal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-crit-dsn.png'
-    ],
-    colorset: 'rage',
-    system: 'vtm5e'
-  })
-
-  dice3d.addColorset({
-    name: 'werewolf',
-    description: 'Werewolf Dice',
-    category: 'V5',
-    foreground: '#000000',
-    background: '#5c6419',
-    texture: 'none',
-    edge: '#5c6419',
-    material: 'plastic',
-    font: 'Arial Black'
-  }, 'default')
-
-  dice3d.addColorset({
-    name: 'rage',
-    description: 'Rage Dice',
-    category: 'V5',
-    foreground: '#000000',
-    background: '#9f0a00',
-    texture: 'none',
-    edge: '#9f0a00',
-    material: 'plastic',
-    font: 'Arial Black'
-  }, 'default')
-})
-
-Hooks.once('diceSoNiceReady', (dice3d) => {
-  dice3d.addSystem({ id: 'vtm5x', name: 'VtM5e Custom' }, true)
-  dice3d.addDicePreset({
-    type: 'dv',
-    labels: [
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-crit-dsn.png'
-    ],
-    system: 'vtm5x'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dg',
-    labels: [
-      'systems/vtm5e/assets/images/bestial-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsn.png',
-      'systems/vtm5e/assets/images/red-crit-dsn.png'
-    ],
-    system: 'vtm5x'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dh',
-    labels: [
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-crit-dsn.png'
-    ],
-    system: 'vtm5x'
-  })
-
-  dice3d.addDicePreset({
-    type: 'ds',
-    labels: [
-      'systems/vtm5e/assets/images/desperation-fail-dsnc.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-crit-dsnc.png'
-    ],
-    system: 'vtm5x'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dw',
-    labels: [
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-crit-dsn.png'
-    ],
-    system: 'vtm5x'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dr',
-    labels: [
-      'systems/vtm5e/assets/images/werewolf-brutal-fail-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-brutal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-success-dsn.png',
-      'systems/vtm5e/assets/images/werewolf-crit-dsn.png'
-    ],
-    system: 'vtm5x'
-  })
-})
-
-Hooks.once('diceSoNiceReady', (dice3d) => {
-  dice3d.addSystem({ id: 'vtm5y', name: 'VtM5e Colors' }, true)
-  dice3d.addDicePreset({
-    type: 'dv',
-    labels: [
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-success-dsn.png',
-      'systems/vtm5e/assets/images/normal-crit-dsn.png'
-    ],
-    system: 'vtm5y'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dg',
-    labels: [
-      'systems/vtm5e/assets/images/bestial-fail-dsnc.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-success-dsnc.png',
-      'systems/vtm5e/assets/images/red-success-dsnc.png',
-      'systems/vtm5e/assets/images/red-success-dsnc.png',
-      'systems/vtm5e/assets/images/red-success-dsnc.png',
-      'systems/vtm5e/assets/images/red-crit-dsnc.png'
-    ],
-    system: 'vtm5y'
-  })
-
-  dice3d.addDicePreset({
-    type: 'dh',
-    labels: [
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/normal-fail-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-crit-dsn.png'
-    ],
-    system: 'vtm5y'
-  })
-
-  dice3d.addDicePreset({
-    type: 'ds',
-    labels: [
-      'systems/vtm5e/assets/images/desperation-fail-dsnc.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/red-fail-dsn.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-success-dsnc.png',
-      'systems/vtm5e/assets/images/hunter-normal-crit-dsnc.png'
-    ],
-    system: 'vtm5y'
-  })
+  loadDiceSoNice(dice3d)
 })
 
 /* -------------------------------------------- */
